+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-TBD
+### Added
+
+- Task `validateManifest`
+  
+  - Check setup of cloud media conversion
 
 ## [3.2.0] 2021-01-15